refactor: added schema_name dsl to cac queries
