+++ crates/context_aware_config/src/api/dimension/utils.rs
@@ -1,10 +1,7 @@
 use crate::helpers::DimensionData;
-use diesel::{
-    r2d2::{ConnectionManager, PooledConnection},
-    PgConnection, RunQueryDsl,
-};
+use diesel::{query_dsl::methods::SchemaNameDsl, RunQueryDsl};
 use jsonschema::{Draft, JSONSchema};
-use service_utils::helpers::extract_dimensions;
+use service_utils::{helpers::extract_dimensions, service::types::Tenant};
 use std::collections::HashMap;
 use superposition_macros::{bad_argument, db_error, unexpected_error};
 use superposition_types::{
@@ -12,15 +9,16 @@ use superposition_types::{
         models::cac::{Context, Dimension},
         schema::{contexts::dsl::contexts, dimensions::dsl::*},
     },
-    result as superposition, Cac, Condition,
+    result as superposition, Cac, Condition, DBConnection,
 };
 
 use super::types::{DimensionName, Position};
 
 pub fn get_dimension_data(
-    conn: &mut PooledConnection<ConnectionManager<PgConnection>>,
+    conn: &mut DBConnection,
+    tenant: &Tenant,
 ) -> superposition::Result<Vec<Dimension>> {
-    Ok(dimensions.load::<Dimension>(conn)?)
+    Ok(dimensions.schema_name(tenant).load::<Dimension>(conn)?)
 }
 
 pub fn get_dimension_data_map(
@@ -49,12 +47,14 @@ pub fn get_dimension_data_map(
 
 pub fn get_dimension_usage_context_ids(
     key: &str,
-    conn: &mut PooledConnection<ConnectionManager<PgConnection>>,
+    conn: &mut DBConnection,
+    tenant: &Tenant,
 ) -> superposition::Result<Vec<String>> {
-    let result: Vec<Context> = contexts.load(conn).map_err(|err| {
-        log::error!("failed to fetch contexts with error: {}", err);
-        db_error!(err)
-    })?;
+    let result: Vec<Context> =
+        contexts.schema_name(tenant).load(conn).map_err(|err| {
+            log::error!("failed to fetch contexts with error: {}", err);
+            db_error!(err)
+        })?;
 
     let mut context_ids = vec![];
     for context in result.iter() {
