+++ crates/context_aware_config/src/api/functions/helpers.rs
@@ -1,10 +1,8 @@
 extern crate base64;
 
 use base64::prelude::*;
-use diesel::{
-    r2d2::{ConnectionManager, PooledConnection},
-    ExpressionMethods, PgConnection, QueryDsl, RunQueryDsl,
-};
+use diesel::{ExpressionMethods, QueryDsl, RunQueryDsl};
+use service_utils::service::types::Tenant;
 use std::str;
 use superposition_macros::unexpected_error;
 use superposition_types::{
@@ -12,15 +10,17 @@ use superposition_types::{
         models::cac::Function,
         schema::{self, functions::dsl::functions},
     },
-    result as superposition,
+    result as superposition, DBConnection,
 };
 
 pub fn fetch_function(
     f_name: &String,
-    conn: &mut PooledConnection<ConnectionManager<PgConnection>>,
+    conn: &mut DBConnection,
+    tenant: &Tenant,
 ) -> superposition::Result<Function> {
     Ok(functions
         .filter(schema::functions::function_name.eq(f_name))
+        .schema_name(tenant)
         .get_result::<Function>(conn)?)
 }
 
@@ -50,19 +50,22 @@ pub fn decode_base64_to_string(code: &String) -> superposition::Result<String> {
 }
 
 pub fn get_published_function_code(
-    conn: &mut PooledConnection<ConnectionManager<PgConnection>>,
+    conn: &mut DBConnection,
     f_name: String,
+    tenant: &Tenant,
 ) -> superposition::Result<Option<String>> {
     let function = functions
         .filter(schema::functions::function_name.eq(f_name))
         .select(schema::functions::published_code)
+        .schema_name(tenant)
         .first(conn)?;
     Ok(function)
 }
 
 pub fn get_published_functions_by_names(
-    conn: &mut PooledConnection<ConnectionManager<PgConnection>>,
+    conn: &mut DBConnection,
     function_names: Vec<String>,
+    tenant: &Tenant,
 ) -> superposition::Result<Vec<(String, Option<String>)>> {
     let function: Vec<(String, Option<String>)> = functions
         .filter(schema::functions::function_name.eq_any(function_names))
@@ -70,6 +73,7 @@ pub fn get_published_functions_by_names(
             schema::functions::function_name,
             schema::functions::published_code,
         ))
+        .schema_name(tenant)
         .load(conn)?;
     Ok(function)
 }
