+++ crates/context_aware_config/src/api/type_templates/handlers.rs
@@ -4,7 +4,7 @@ use chrono::Utc;
 use diesel::{ExpressionMethods, OptionalExtension, QueryDsl, RunQueryDsl};
 use jsonschema::JSONSchema;
 use serde_json::Value;
-use service_utils::service::types::DbConnection;
+use service_utils::service::types::{DbConnection, Tenant};
 use superposition_macros::{bad_argument, db_error};
 use superposition_types::{
     custom_query::PaginationParams,
@@ -30,6 +30,7 @@ async fn create_type(
     request: Json<TypeTemplateCreateRequest>,
     db_conn: DbConnection,
     user: User,
+    tenant: Tenant,
 ) -> superposition::Result<HttpResponse> {
     let DbConnection(mut conn) = db_conn;
     let _ = JSONSchema::compile(&request.type_schema).map_err(|err| {
@@ -53,6 +54,7 @@ async fn create_type(
             type_templates::description.eq(request.description.clone()),
             type_templates::change_reason.eq(request.change_reason.clone()),
         ))
+        .schema_name(&tenant)
         .get_result::<TypeTemplate>(&mut conn)
         .map_err(|err| {
             log::error!("failed to insert custom type with error: {}", err);
@@ -67,6 +69,7 @@ async fn update_type(
     path: Path<TypeTemplateName>,
     db_conn: DbConnection,
     user: User,
+    tenant: Tenant,
 ) -> superposition::Result<HttpResponse> {
     let DbConnection(mut conn) = db_conn;
     let _ = JSONSchema::compile(&request).map_err(|err| {
@@ -114,6 +117,7 @@ async fn update_type(
             type_templates::last_modified_by.eq(user.email),
             type_templates::description.eq(final_description),
         ))
+        .schema_name(&tenant)
         .get_result::<TypeTemplate>(&mut conn)
         .map_err(|err| {
             log::error!("failed to insert custom type with error: {}", err);
@@ -127,6 +131,7 @@ async fn delete_type(
     path: Path<TypeTemplateName>,
     db_conn: DbConnection,
     user: User,
+    tenant: Tenant,
 ) -> superposition::Result<HttpResponse> {
     let DbConnection(mut conn) = db_conn;
     let type_name: String = path.into_inner().into();
@@ -136,9 +141,11 @@ async fn delete_type(
             dsl::last_modified_at.eq(Utc::now().naive_utc()),
             dsl::last_modified_by.eq(user.email),
         ))
+        .schema_name(&tenant)
         .execute(&mut conn)?;
     let deleted_type =
         diesel::delete(dsl::type_templates.filter(dsl::type_name.eq(type_name)))
+            .schema_name(&tenant)
             .get_result::<TypeTemplate>(&mut conn)?;
     Ok(HttpResponse::Ok().json(deleted_type))
 }
@@ -147,12 +154,14 @@ async fn delete_type(
 async fn list_types(
     db_conn: DbConnection,
     filters: Query<PaginationParams>,
+    tenant: Tenant,
 ) -> superposition::Result<Json<PaginatedResponse<TypeTemplate>>> {
     let DbConnection(mut conn) = db_conn;
 
     if let Some(true) = filters.all {
-        let result: Vec<TypeTemplate> =
-            type_templates::dsl::type_templates.get_results(&mut conn)?;
+        let result: Vec<TypeTemplate> = type_templates::dsl::type_templates
+            .schema_name(&tenant)
+            .get_results(&mut conn)?;
         return Ok(Json(PaginatedResponse {
             total_pages: 1,
             total_items: result.len() as i64,
@@ -162,12 +171,14 @@ async fn list_types(
 
     let n_types: i64 = type_templates::dsl::type_templates
         .count()
+        .schema_name(&tenant)
         .get_result(&mut conn)?;
     let limit = filters.count.unwrap_or(10);
     let mut builder = type_templates::dsl::type_templates
-        .into_boxed()
+        .schema_name(&tenant)
         .order(type_templates::dsl::created_at.desc())
-        .limit(limit);
+        .limit(limit)
+        .into_boxed();
     if let Some(page) = filters.page {
         let offset = (page - 1) * limit;
         builder = builder.offset(offset);
