+++ crates/context_aware_config/src/helpers.rs
@@ -6,8 +6,7 @@ use actix_web::web::Data;
 use chrono::DateTime;
 use chrono::Utc;
 use diesel::{
-    r2d2::{ConnectionManager, PooledConnection},
-    ExpressionMethods, PgConnection, QueryDsl, RunQueryDsl,
+    ExpressionMethods, QueryDsl, RunQueryDsl,
 };
 #[cfg(feature = "high-performance-mode")]
 use fred::interfaces::KeysInterface;
@@ -17,6 +16,7 @@ use jsonlogic;
 use jsonschema::{Draft, JSONSchema, ValidationError};
 use num_bigint::BigUint;
 use serde_json::{json, Map, Value};
+use service_utils::service::types::Tenant;
 #[cfg(feature = "high-performance-mode")]
 use service_utils::service::types::Tenant;
 use service_utils::{
@@ -25,7 +25,8 @@ use service_utils::{
 };
 use superposition_macros::{bad_argument, db_error, unexpected_error, validation_error};
 #[cfg(feature = "high-performance-mode")]
-use superposition_types::database::schema::event_log::dsl as event_log;
+use superposition_types::cac::schema::event_log::dsl as event_log;
+use superposition_types::DBConnection;
 use superposition_types::{
     database::{
         models::cac::ConfigVersion,
@@ -215,11 +216,13 @@ pub fn calculate_context_weight(
     Ok(weight)
 }
 pub fn generate_cac(
-    conn: &mut PooledConnection<ConnectionManager<PgConnection>>,
+    conn: &mut DBConnection,
+    tenant: &Tenant,
 ) -> superposition::Result<Config> {
     let contexts_vec: Vec<(String, Condition, String, Overrides)> = ctxt::contexts
         .select((ctxt::id, ctxt::value, ctxt::override_id, ctxt::override_))
         .order_by((ctxt::weight.asc(), ctxt::created_at.asc()))
+        .schema_name(tenant)
         .load::<(String, Condition, String, Overrides)>(conn)
         .map_err(|err| {
             log::error!("failed to fetch contexts with error: {}", err);
@@ -269,6 +272,7 @@ pub fn generate_cac(
 
     let default_config_vec = def_conf::default_configs
         .select((def_conf::key, def_conf::value))
+        .schema_name(&tenant)
         .load::<(String, Value)>(conn)
         .map_err(|err| {
             log::error!("failed to fetch default_configs with error: {}", err);
@@ -293,13 +297,14 @@ pub fn generate_cac(
 pub fn add_config_version(
     state: &Data<AppState>,
     tags: Option<Vec<String>>,
-    db_conn: &mut PooledConnection<ConnectionManager<PgConnection>>,
     description: String,
     change_reason: String,
+    db_conn: &mut DBConnection,
+    tenant: &Tenant,
 ) -> superposition::Result<i64> {
     use config_versions::dsl::config_versions;
     let version_id = generate_snowflake_id(state)?;
-    let config = generate_cac(db_conn)?;
+    let config = generate_cac(db_conn, tenant)?;
     let json_config = json!(config);
     let config_hash = blake3::hash(json_config.to_string().as_bytes()).to_string();
     let config_version = ConfigVersion {
@@ -313,6 +318,7 @@ pub fn add_config_version(
     };
     diesel::insert_into(config_versions)
         .values(&config_version)
+        .schema_name(tenant)
         .execute(db_conn)?;
     Ok(version_id)
 }
@@ -324,7 +330,7 @@ pub async fn put_config_in_redis(
     tenant: Tenant,
     db_conn: &mut PooledConnection<ConnectionManager<PgConnection>>,
 ) -> superposition::Result<()> {
-    let raw_config = generate_cac(db_conn)?;
+    let raw_config = generate_cac(db_conn, &tenant)?;
     let parsed_config = serde_json::to_string(&json!(raw_config)).map_err(|e| {
         log::error!("failed to convert cac config to string: {}", e);
         unexpected_error!("could not convert cac config to string")
