+++ crates/superposition_types/src/lib.rs
@@ -13,7 +13,9 @@ use std::fmt::Display;
 use std::future::{ready, Ready};
 
 #[cfg(feature = "server")]
-use actix_web::{dev::Payload, FromRequest, HttpMessage, HttpRequest};
+use actix_web::{dev::Payload, error, FromRequest, HttpMessage, HttpRequest};
+use diesel::r2d2::{ConnectionManager, PooledConnection};
+use diesel::PgConnection;
 #[cfg(feature = "diesel_derives")]
 use diesel_derive_enum as _;
 use regex::Regex;
@@ -208,6 +210,8 @@ impl Default for SortBy {
         Self::Desc
     }
 }
+
+pub type DBConnection = PooledConnection<ConnectionManager<PgConnection>>;
 
 #[cfg(test)]
 mod tests {
